int, double -> float

+++ app/src/main/java/com/example/youtubecleaner/DemoData.java
@@ -7,7 +7,7 @@ public class DemoData {
     private String[] arrComment = {"역시 재미있어~ 다음 컨텐츠도 기대된다 ㅋㅋ", "구독하고 열심히 정주행 해야겠다 ㅎㅎ", "사람 비꼬면서 하는 개그 재밌나 ㅜㅜ 진짜 예의없다",
             "오랜만에 이렇게 웃어보네요\n너무 재미있어요\n항상 웃는 일로 가득했으면~", "아진짜 너무 귀엽다 ㅋㅋㅋㅋㅋㅋ", "아 저 매력 어쩌냐고 ㅠㅠㅋㅋㅋㅋ", "아 진짜 레전드다 ㅋㅋㅋㅋㅋㅋㅋㅋ", "으 더러워... 저게 뭐가 재밌누",
             "으... 듣기 거북함....", "이번 편은 별로 재미 없는듯? 기대이하네", "이 영상 뭔가 중독적... 5번은 본 듯", "영상 하나하나가 소중해 ㅜㅜㅜ", "레전드 개웃김 ㅋㅋㅋㅋㅋㅋ 쿵짝 맞는거봐 ㄹㅇㅋㅋㅋㅋㅋ"};
-    private Integer[] arrScore = {10, 1, 72, 1, 10, 1, 15, 86, 74, 66, 28, 8, 10};
+    private Float[] arrScore = {10.0f, 1.0f, 72.0f, 1.0f, 10.0f, 1.0f, 15.0f, 86.0f, 74.0f, 66.0f, 28.0f, 8.0f, 10.0f};
 
     int numComment = arrUserID.length;
 
@@ -19,7 +19,7 @@ public class DemoData {
         return arrComment;
     }
 
-    public Integer[] getScore(){
+    public Float[] getScore(){
         return arrScore;
     }
